Added feature to disable temp directory deletion in response to issue #3

+++ django_drf_filepond/models.py
@@ -10,6 +10,8 @@ from django.db.models.signals import post_delete
 from django.dispatch import receiver
 from django.core.files.storage import FileSystemStorage
 
+import django_drf_filepond.drf_filepond_settings as local_settings
+
 FILEPOND_UPLOAD_TMP = getattr(settings, 'DJANGO_DRF_FILEPOND_UPLOAD_TMP',
                         os.path.join(settings.BASE_DIR,'filepond_uploads'))
 storage = FileSystemStorage(location=FILEPOND_UPLOAD_TMP)
@@ -56,8 +58,10 @@ def delete_temp_upload_file(sender, instance, **kwargs):
             os.path.isfile(instance.file.path)):
             os.remove(instance.file.path)
     
-    file_dir = os.path.join(storage.location, instance.upload_id)
-    if(os.path.exists(file_dir) and os.path.isdir(file_dir)):
-        os.rmdir(file_dir)
-        LOG.debug('*** post_delete signal handler called. Deleting temp dir that contained file.')
+    if local_settings.DELETE_UPLOAD_TMP_DIRS:
+        file_dir = os.path.join(storage.location, instance.upload_id)
+        if(os.path.exists(file_dir) and os.path.isdir(file_dir)):
+            os.rmdir(file_dir)
+            LOG.debug('*** post_delete signal handler called. Deleting temp '
+                      'dir that contained file.')
